+++ apps/example-expo/app/example-above-bottom-sheet-view.tsx
@@ -1,5 +1,6 @@
+import { ThemedText } from '@/components/text'
 import React, { Fragment, useState } from 'react'
-import { Button, StyleSheet, Text, View } from 'react-native'
+import { Button, StyleSheet, View } from 'react-native'
 import Animated, { useAnimatedStyle } from 'react-native-reanimated'
 import {
   BottomSheet,
@@ -32,7 +33,7 @@ export default function ExampleAboveBottomSheetView() {
     return (
       <Fragment>
         {Array.from({ length: 20 }).map((_, index) => (
-          <Text key={index}>Item {index + 1}</Text>
+          <ThemedText key={index}>Item {index + 1}</ThemedText>
         ))}
       </Fragment>
     )
@@ -40,7 +41,9 @@ export default function ExampleAboveBottomSheetView() {
 
   return (
     <View style={styles.root}>
-      <Text style={styles.header}>Example Above Bottom Sheet View</Text>
+      <ThemedText style={styles.header}>
+        Example Above Bottom Sheet View
+      </ThemedText>
 
       <Button title="Open Sheet A" onPress={() => setIsOpenA(true)} />
 
@@ -65,11 +68,13 @@ export default function ExampleAboveBottomSheetView() {
                 <BottomSheetHandle />
 
                 <BottomSheetView>
-                  <Text>Sheet A</Text>
+                  <ThemedText>Sheet A</ThemedText>
+
                   <Button
                     title="Close Sheet A"
                     onPress={() => setIsOpenA(false)}
                   />
+
                   {renderContent()}
                 </BottomSheetView>
               </BottomSheet>