+++ apps/example-expo/app/example-bottom-sheet-float-mode.tsx
@@ -1,5 +1,6 @@
+import { ThemedText } from '@/components/text'
 import { Fragment, useState } from 'react'
-import { Button, StyleSheet, Text, View } from 'react-native'
+import { Button, StyleSheet, View } from 'react-native'
 import {
   Backdrop,
   BottomSheet,
@@ -19,7 +20,7 @@ export default function ExampleBottomSheetFloatMode() {
     return (
       <Fragment>
         {Array.from({ length: 20 }).map((_, index) => (
-          <Text key={index}>Item {index + 1}</Text>
+          <ThemedText key={index}>Item {index + 1}</ThemedText>
         ))}
       </Fragment>
     )
@@ -27,7 +28,9 @@ export default function ExampleBottomSheetFloatMode() {
 
   return (
     <View style={styles.root}>
-      <Text style={styles.header}>Example Bottom Sheet (Float Mode)</Text>
+      <ThemedText style={styles.header}>
+        Example Bottom Sheet (Float Mode)
+      </ThemedText>
 
       <Button
         title="Open Sheet A (Dynamic sizing)"
@@ -54,11 +57,13 @@ export default function ExampleBottomSheetFloatMode() {
                 <BottomSheetHandle />
 
                 <BottomSheetView>
-                  <Text>Sheet A</Text>
+                  <ThemedText>Sheet A</ThemedText>
+
                   <Button
                     title="Close Sheet A"
                     onPress={() => setIsOpenA(false)}
                   />
+
                   {renderContent()}
                 </BottomSheetView>
               </BottomSheet>
@@ -82,11 +87,13 @@ export default function ExampleBottomSheetFloatMode() {
                 <BottomSheetHandle />
 
                 <BottomSheetView>
-                  <Text>Sheet B</Text>
+                  <ThemedText>Sheet B</ThemedText>
+
                   <Button
                     title="Close Sheet B"
                     onPress={() => setIsOpenB(false)}
                   />
+
                   {renderContent()}
                 </BottomSheetView>
               </BottomSheet>