+++ packages/react-native-the-sheet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # react-native-the-sheet
 
+## 2.0.14
+
+### Patch Changes
+
+- Make `AnimatedProp` + `SnapPoint` types public
+
 ## 2.0.13
 
 ### Patch Changes